Split legacy Mimir receiver config out of PostableApiReceiver
PostableApiReceiver now only carries Grafana-native receiver config
(Name + GrafanaManagedReceivers). The legacy Mimir integration struct
moved to definition/compat as compat.Receiver. LoadCompat no longer
applies per-integration Mimir defaults since PostableApiReceiver can't
hold that data. PostableMimirReceiverToPostableGrafanaReceiver now takes
a compat.Receiver directly and returns a fresh PostableApiReceiver.
