+++ definition/alertmanager.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"reflect"
 	"sort"
 	"strings"
 
@@ -14,22 +13,6 @@ import (
 	"github.com/prometheus/alertmanager/config"
 	"github.com/prometheus/alertmanager/pkg/labels"
 	"github.com/prometheus/common/model"
-
-	discord_v0mimir1 "github.com/grafana/alerting/receivers/discord/v0mimir1"
-	email_v0mimir1 "github.com/grafana/alerting/receivers/email/v0mimir1"
-	jira_v0mimir1 "github.com/grafana/alerting/receivers/jira/v0mimir1"
-	opsgenie_v0mimir1 "github.com/grafana/alerting/receivers/opsgenie/v0mimir1"
-	pagerduty_v0mimir1 "github.com/grafana/alerting/receivers/pagerduty/v0mimir1"
-	pushover_v0mimir1 "github.com/grafana/alerting/receivers/pushover/v0mimir1"
-	slack_v0mimir1 "github.com/grafana/alerting/receivers/slack/v0mimir1"
-	sns_v0mimir1 "github.com/grafana/alerting/receivers/sns/v0mimir1"
-	teams_v0mimir1 "github.com/grafana/alerting/receivers/teams/v0mimir1"
-	teams_v0mimir2 "github.com/grafana/alerting/receivers/teams/v0mimir2"
-	telegram_v0mimir1 "github.com/grafana/alerting/receivers/telegram/v0mimir1"
-	victorops_v0mimir1 "github.com/grafana/alerting/receivers/victorops/v0mimir1"
-	webex_v0mimir1 "github.com/grafana/alerting/receivers/webex/v0mimir1"
-	webhook_v0mimir1 "github.com/grafana/alerting/receivers/webhook/v0mimir1"
-	wechat_v0mimir1 "github.com/grafana/alerting/receivers/wechat/v0mimir1"
 )
 
 type Provenance string
@@ -516,15 +499,14 @@ func (m ObjectMatchers) MarshalJSON() ([]byte, error) {
 
 // nolint:revive
 type PostableApiReceiver struct {
-	Receiver                 `yaml:",inline"`
+	Name                     string `yaml:"name" json:"name"`
 	PostableGrafanaReceivers `yaml:",inline"`
 }
 
 func (r *PostableApiReceiver) Validate() error {
 	var errs []error
-	err := r.Receiver.Validate()
-	if err != nil {
-		errs = append(errs, err)
+	if r.Name == "" {
+		errs = append(errs, errors.New("missing name in receiver"))
 	}
 	return errors.Join(errs...)
 }
@@ -534,22 +516,8 @@ func (r *PostableApiReceiver) UnmarshalJSON(b []byte) error {
 }
 
 func (r *PostableApiReceiver) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	if err := unmarshal(&r.PostableGrafanaReceivers); err != nil {
-		return err
-	}
-
-	type plain Receiver
-	if err := unmarshal((*plain)(&r.Receiver)); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (r *PostableApiReceiver) HasMimirIntegrations() bool {
-	cpy := r.Receiver
-	cpy.Name = ""
-	return !reflect.ValueOf(cpy).IsZero()
+	type plain PostableApiReceiver
+	return unmarshal((*plain)(r))
 }
 
 func (r *PostableApiReceiver) HasGrafanaIntegrations() bool {
@@ -565,33 +533,6 @@ func CopyIntegrations(src, dest *PostableApiReceiver) error {
 	if src == nil || dest == nil {
 		return errors.New("both source and destination receivers should be non-nil")
 	}
-	// Get the reflect.Value of src and dest
-	srcVal := reflect.ValueOf(&src.Receiver).Elem()
-	destVal := reflect.ValueOf(&dest.Receiver).Elem()
-
-	// Iterate through all fields of the struct
-	for i := 0; i < srcVal.NumField(); i++ {
-		srcField := srcVal.Field(i)
-		destField := destVal.Field(i)
-
-		// Only process slice fields (skip Name field)
-		if srcField.Kind() != reflect.Slice {
-			continue
-		}
-
-		// Get the length of the source slice
-		srcLen := srcField.Len()
-
-		// Append each element from source to destination
-		for j := 0; j < srcLen; j++ {
-			// Get the element from source slice
-			elem := srcField.Index(j)
-
-			// Append to destination slice
-			destField.Set(reflect.Append(destField, elem))
-		}
-	}
-
 	if src.GrafanaManagedReceivers != nil {
 		dest.GrafanaManagedReceivers = append(dest.GrafanaManagedReceivers, src.GrafanaManagedReceivers...)
 	}
@@ -649,124 +590,3 @@ type TemplateKind string
 
 const GrafanaTemplateKind TemplateKind = "grafana"
 const MimirTemplateKind TemplateKind = "mimir"
-
-// Receiver configuration provides configuration on how to contact a receiver.
-type Receiver struct {
-	// A unique identifier for this receiver.
-	Name string `yaml:"name" json:"name"`
-
-	DiscordConfigs   []*discord_v0mimir1.Config   `yaml:"discord_configs,omitempty" json:"discord_configs,omitempty"`
-	EmailConfigs     []*email_v0mimir1.Config     `yaml:"email_configs,omitempty" json:"email_configs,omitempty"`
-	PagerdutyConfigs []*pagerduty_v0mimir1.Config `yaml:"pagerduty_configs,omitempty" json:"pagerduty_configs,omitempty"`
-	SlackConfigs     []*slack_v0mimir1.Config     `yaml:"slack_configs,omitempty" json:"slack_configs,omitempty"`
-	WebhookConfigs   []*webhook_v0mimir1.Config   `yaml:"webhook_configs,omitempty" json:"webhook_configs,omitempty"`
-	OpsGenieConfigs  []*opsgenie_v0mimir1.Config  `yaml:"opsgenie_configs,omitempty" json:"opsgenie_configs,omitempty"`
-	WechatConfigs    []*wechat_v0mimir1.Config    `yaml:"wechat_configs,omitempty" json:"wechat_configs,omitempty"`
-	PushoverConfigs  []*pushover_v0mimir1.Config  `yaml:"pushover_configs,omitempty" json:"pushover_configs,omitempty"`
-	VictorOpsConfigs []*victorops_v0mimir1.Config `yaml:"victorops_configs,omitempty" json:"victorops_configs,omitempty"`
-	SNSConfigs       []*sns_v0mimir1.Config       `yaml:"sns_configs,omitempty" json:"sns_configs,omitempty"`
-	TelegramConfigs  []*telegram_v0mimir1.Config  `yaml:"telegram_configs,omitempty" json:"telegram_configs,omitempty"`
-	WebexConfigs     []*webex_v0mimir1.Config     `yaml:"webex_configs,omitempty" json:"webex_configs,omitempty"`
-	MSTeamsConfigs   []*teams_v0mimir1.Config     `yaml:"msteams_configs,omitempty" json:"msteams_configs,omitempty"`
-	MSTeamsV2Configs []*teams_v0mimir2.Config     `yaml:"msteamsv2_configs,omitempty" json:"msteamsv2_configs,omitempty"`
-	JiraConfigs      []*jira_v0mimir1.Config      `yaml:"jira_configs,omitempty" json:"jira_configs,omitempty"`
-}
-
-// UnmarshalYAML implements the yaml.Unmarshaler interface for Receiver.
-func (c *Receiver) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	type plain Receiver
-	if err := unmarshal((*plain)(c)); err != nil {
-		return err
-	}
-	if c.Name == "" {
-		return fmt.Errorf("missing name in receiver")
-	}
-	return nil
-}
-
-// Validate calls Validate on all integration configs and accumulates errors.
-// Each error is annotated with the integration type (JSON tag) and index.
-func (c *Receiver) Validate() error {
-	var errs []error
-
-	if c.Name == "" {
-		errs = append(errs, errors.New("missing name in receiver"))
-	}
-
-	for i, cfg := range c.DiscordConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("discord [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.EmailConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("email [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.PagerdutyConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("pagerduty [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.SlackConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("slack [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.WebhookConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("webhook [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.OpsGenieConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("opsgenie [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.WechatConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("wechat [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.PushoverConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("pushover [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.VictorOpsConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("victorops [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.SNSConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("sns [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.TelegramConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("telegram [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.WebexConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("webex [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.MSTeamsConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("msteams [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.MSTeamsV2Configs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("msteamsv2 [%d]: %w", i, err))
-		}
-	}
-	for i, cfg := range c.JiraConfigs {
-		if err := cfg.Validate(); err != nil {
-			errs = append(errs, fmt.Errorf("jira [%d]: %w", i, err))
-		}
-	}
-	return errors.Join(errs...)
-}
