+++ definition/compat.go
@@ -10,9 +10,6 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/prometheus/alertmanager/config"
-
-	"github.com/grafana/alerting/http/v0mimir"
-	"github.com/grafana/alerting/receivers"
 )
 
 // LoadCompat loads a PostableApiAlertingConfig from a YAML configuration
@@ -33,206 +30,11 @@ func LoadCompat(rawCfg []byte) (*PostableApiAlertingConfig, error) {
 		*c.Global = config.DefaultGlobalConfig()
 	}
 
-	// Check that the configuration for upstream receivers is well formed.
-	// Taken from https://github.com/prometheus/alertmanager/blob/3e70148d4f55a27b0c817d0997149bff30f6343e/config/config.go#L404-L637
-	// with some modifications to ignore file-related settings and exclude RocketChat (not supported yet.
+	// Check that receiver names are unique.
 	names := map[string]struct{}{}
 	for _, rcv := range c.Receivers {
 		if _, ok := names[rcv.Name]; ok {
 			return nil, fmt.Errorf("notification config name %q is not unique", rcv.Name)
-		}
-		for _, wh := range rcv.WebhookConfigs {
-			if wh.HTTPConfig == nil {
-				wh.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-		}
-		for _, ec := range rcv.EmailConfigs {
-			if ec.Smarthost.String() == "" {
-				if c.Global.SMTPSmarthost.String() == "" {
-					return nil, errors.New("no global SMTP smarthost set")
-				}
-				ec.Smarthost = receivers.HostPort(c.Global.SMTPSmarthost)
-			}
-			if ec.From == "" {
-				if c.Global.SMTPFrom == "" {
-					return nil, errors.New("no global SMTP from set")
-				}
-				ec.From = c.Global.SMTPFrom
-			}
-			if ec.Hello == "" {
-				ec.Hello = c.Global.SMTPHello
-			}
-			if ec.AuthUsername == "" {
-				ec.AuthUsername = c.Global.SMTPAuthUsername
-			}
-			if ec.AuthPassword == "" {
-				ec.AuthPassword = receivers.Secret(c.Global.SMTPAuthPassword)
-			}
-			if ec.AuthSecret == "" {
-				ec.AuthSecret = receivers.Secret(c.Global.SMTPAuthSecret)
-			}
-			if ec.AuthIdentity == "" {
-				ec.AuthIdentity = c.Global.SMTPAuthIdentity
-			}
-			if ec.RequireTLS == nil {
-				ec.RequireTLS = new(bool)
-				*ec.RequireTLS = c.Global.SMTPRequireTLS
-			}
-		}
-		for _, sc := range rcv.SlackConfigs {
-			if sc.HTTPConfig == nil {
-				sc.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-			if sc.APIURL == nil {
-				if c.Global.SlackAPIURL == nil {
-					return nil, errors.New("no global Slack API URL set")
-				}
-				sc.APIURL = (*receivers.SecretURL)(c.Global.SlackAPIURL)
-			}
-		}
-		for _, poc := range rcv.PushoverConfigs {
-			if poc.HTTPConfig == nil {
-				poc.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-		}
-		for _, pdc := range rcv.PagerdutyConfigs {
-			if pdc.HTTPConfig == nil {
-				pdc.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-			if pdc.URL == nil {
-				if c.Global.PagerdutyURL == nil {
-					return nil, errors.New("no global PagerDuty URL set")
-				}
-				pdc.URL = (*receivers.URL)(c.Global.PagerdutyURL)
-			}
-		}
-		for _, ogc := range rcv.OpsGenieConfigs {
-			if ogc.HTTPConfig == nil {
-				ogc.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-			if ogc.APIURL == nil {
-				if c.Global.OpsGenieAPIURL == nil {
-					return nil, errors.New("no global OpsGenie URL set")
-				}
-				ogc.APIURL = (*receivers.URL)(c.Global.OpsGenieAPIURL)
-			}
-			if !strings.HasSuffix(ogc.APIURL.Path, "/") {
-				ogc.APIURL.Path += "/"
-			}
-			if ogc.APIKey == "" {
-				if c.Global.OpsGenieAPIKey == "" {
-					return nil, errors.New("no global OpsGenie API Key set")
-				}
-				ogc.APIKey = receivers.Secret(c.Global.OpsGenieAPIKey)
-			}
-		}
-		for _, wcc := range rcv.WechatConfigs {
-			if wcc.HTTPConfig == nil {
-				wcc.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-
-			if wcc.APIURL == nil {
-				if c.Global.WeChatAPIURL == nil {
-					return nil, errors.New("no global Wechat URL set")
-				}
-				wcc.APIURL = (*receivers.URL)(c.Global.WeChatAPIURL)
-			}
-
-			if wcc.APISecret == "" {
-				if c.Global.WeChatAPISecret == "" {
-					return nil, errors.New("no global Wechat ApiSecret set")
-				}
-				wcc.APISecret = receivers.Secret(c.Global.WeChatAPISecret)
-			}
-
-			if wcc.CorpID == "" {
-				if c.Global.WeChatAPICorpID == "" {
-					return nil, errors.New("no global Wechat CorpID set")
-				}
-				wcc.CorpID = c.Global.WeChatAPICorpID
-			}
-
-			if !strings.HasSuffix(wcc.APIURL.Path, "/") {
-				wcc.APIURL.Path += "/"
-			}
-		}
-		for _, voc := range rcv.VictorOpsConfigs {
-			if voc.HTTPConfig == nil {
-				voc.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-			if voc.APIURL == nil {
-				if c.Global.VictorOpsAPIURL == nil {
-					return nil, errors.New("no global VictorOps URL set")
-				}
-				voc.APIURL = (*receivers.URL)(c.Global.VictorOpsAPIURL)
-			}
-			if !strings.HasSuffix(voc.APIURL.Path, "/") {
-				voc.APIURL.Path += "/"
-			}
-			if voc.APIKey == "" {
-				if c.Global.VictorOpsAPIKey == "" {
-					return nil, errors.New("no global VictorOps API Key set")
-				}
-				voc.APIKey = receivers.Secret(c.Global.VictorOpsAPIKey)
-			}
-		}
-		for _, sns := range rcv.SNSConfigs {
-			if sns.HTTPConfig == nil {
-				sns.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-		}
-
-		for _, telegram := range rcv.TelegramConfigs {
-			if telegram.HTTPConfig == nil {
-				telegram.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-			if telegram.APIUrl == nil {
-				telegram.APIUrl = (*receivers.URL)(c.Global.TelegramAPIUrl)
-			}
-		}
-		for _, discord := range rcv.DiscordConfigs {
-			if discord.HTTPConfig == nil {
-				discord.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-			if discord.WebhookURL == nil {
-				return nil, errors.New("no discord webhook URL provided")
-			}
-		}
-		for _, webex := range rcv.WebexConfigs {
-			if webex.APIURL == nil {
-				if c.Global.WebexAPIURL == nil {
-					return nil, errors.New("no global Webex URL set")
-				}
-
-				webex.APIURL = (*receivers.URL)(c.Global.WebexAPIURL)
-			}
-		}
-		for _, msteams := range rcv.MSTeamsConfigs {
-			if msteams.HTTPConfig == nil {
-				msteams.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-			if msteams.WebhookURL == nil {
-				return nil, errors.New("no msteams webhook URL provided")
-			}
-		}
-		for _, msteamsv2 := range rcv.MSTeamsV2Configs {
-			if msteamsv2.HTTPConfig == nil {
-				msteamsv2.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-			if msteamsv2.WebhookURL == nil {
-				return nil, errors.New("no msteamsv2 webhook URL provided")
-			}
-		}
-		for _, jira := range rcv.JiraConfigs {
-			if jira.HTTPConfig == nil {
-				jira.HTTPConfig = v0mimir.FromCommonHTTPClientConfig(c.Global.HTTPConfig)
-			}
-			if jira.APIURL == nil {
-				if c.Global.JiraAPIURL == nil {
-					return nil, errors.New("no global Jira Cloud URL set")
-				}
-				jira.APIURL = (*receivers.URL)(c.Global.JiraAPIURL)
-			}
 		}
 		names[rcv.Name] = struct{}{}
 	}
