+++ notify/compat.go
@@ -8,6 +8,7 @@ import (
 	"slices"
 
 	"github.com/grafana/alerting/definition"
+	"github.com/grafana/alerting/definition/compat"
 	"github.com/grafana/alerting/models"
 	"github.com/grafana/alerting/receivers/schema"
 	"github.com/grafana/alerting/templates"
@@ -64,32 +65,25 @@ func PostableGrafanaReceiverToIntegrationConfig(r *definition.PostableGrafanaRec
 	}, nil
 }
 
-// PostableMimirReceiverToPostableGrafanaReceiver converts all legacy models to apimodels.PostableGrafanaReceiver.
-// If receiver does not have any legacy receivers, returns the original receiver.
-// Otherwise, returns a copy that contains converted integrations (and shallow copy of existing Grafana integrations).
-func PostableMimirReceiverToPostableGrafanaReceiver(r *definition.PostableApiReceiver) (*definition.PostableApiReceiver, error) {
-	if !r.HasMimirIntegrations() {
-		return r, nil
-	}
-	v0, err := ConfigReceiverToMimirIntegrations(r.Receiver)
+// PostableMimirReceiverToPostableGrafanaReceiver converts legacy Mimir integrations from legacy into
+// a PostableApiReceiver with equivalent Grafana-native integrations.
+func PostableMimirReceiverToPostableGrafanaReceiver(legacy compat.Receiver) (*definition.PostableApiReceiver, error) {
+	v0, err := ConfigReceiverToMimirIntegrations(legacy)
 	if err != nil {
 		return nil, fmt.Errorf("failed to convert v0 receiver to integrations: %w", err)
 	}
 	result := &definition.PostableApiReceiver{
-		Receiver: definition.Receiver{
-			Name: r.Name,
-		},
+		Name: legacy.Name,
 		PostableGrafanaReceivers: definition.PostableGrafanaReceivers{
-			GrafanaManagedReceivers: make([]*definition.PostableGrafanaReceiver, 0, len(v0)+len(r.GrafanaManagedReceivers)),
+			GrafanaManagedReceivers: make([]*definition.PostableGrafanaReceiver, 0, len(v0)),
 		},
 	}
-	result.GrafanaManagedReceivers = append(result.GrafanaManagedReceivers, r.GrafanaManagedReceivers...)
 	typeCount := make(map[string]int)
 	for _, config := range v0 {
 		integrationType := string(config.Schema.Type())
 		idx := typeCount[integrationType]
 		typeCount[integrationType]++
-		integration, err := MimirIntegrationConfigToPostableGrafanaReceiver(config, r.Name, idx)
+		integration, err := MimirIntegrationConfigToPostableGrafanaReceiver(config, legacy.Name, idx)
 		if err != nil {
 			return nil, fmt.Errorf("failed to convert Mimir integration config to PostableGrafanaReceiver: %w", err)
 		}
@@ -142,8 +136,8 @@ func PostableAPITemplatesToTemplateDefinitions(ts []definition.PostableApiTempla
 	return defs
 }
 
-// ConfigReceiverToMimirIntegrations converts a ConfigReceiver to a list of MimirIntegrationConfig
-func ConfigReceiverToMimirIntegrations(receiver ConfigReceiver) ([]MimirIntegrationConfig, error) {
+// ConfigReceiverToMimirIntegrations converts a compat.Receiver to a list of MimirIntegrationConfig
+func ConfigReceiverToMimirIntegrations(receiver compat.Receiver) ([]MimirIntegrationConfig, error) {
 	result := make([]MimirIntegrationConfig, 0)
 	receiverVal := reflect.ValueOf(&receiver).Elem()
 	receiverType := receiverVal.Type()
